+++ Basic/2DSpaceShooter/Assets/Scripts/Spawner.cs
@@ -1,4 +1,4 @@
-﻿using MLAPI;
+﻿using Unity.Netcode;
 using UnityEngine;
 
 public class Spawner : MonoBehaviour
@@ -69,7 +69,7 @@ public class Spawner : MonoBehaviour
     {
         for (int i = 0; i < m_Amount; i++)
         {
-            GameObject go = m_ObjectPool.GetNetworkObject(m_AsteroidPrefab);
+            GameObject go = m_ObjectPool.GetNetworkObject(m_AsteroidPrefab).gameObject;
             go.transform.position = new Vector3(Random.Range(-40, 40), Random.Range(-40, 40));
 
             go.transform.localScale = new Vector3(4, 4, 4);
@@ -81,11 +81,11 @@ public class Spawner : MonoBehaviour
             go.transform.rotation = Quaternion.Euler(0, 0, dir);
             go.GetComponent<Rigidbody2D>().angularVelocity = dir;
             go.GetComponent<Rigidbody2D>().velocity = new Vector2(dx, dy);
-
-            go.GetComponent<NetworkObject>().Spawn(null, true);
+            go.GetComponent<Asteroid>().asteroidPrefab = m_AsteroidPrefab;
+            go.GetComponent<NetworkObject>().Spawn( true); // TODO
         }
     }
-
+    
     void SpawnObstacles()
     {
         // Obstacles are not networked we just spawn them as static objects on each peer
@@ -128,10 +128,10 @@ public class Spawner : MonoBehaviour
                 hits = Physics2D.OverlapCircleAll(pos, 2.0f);
             }
 
-            GameObject powerUp = m_ObjectPool.GetNetworkObject(m_PowerupPrefab);
+            GameObject powerUp = m_ObjectPool.GetNetworkObject(m_PowerupPrefab).gameObject;
             powerUp.transform.position = pos;
             powerUp.GetComponent<Powerup>().buffType.Value = (Buff.BuffType)Random.Range(0, (int)Buff.BuffType.Last);
-            powerUp.GetComponent<NetworkObject>().Spawn(null, true);
+            powerUp.GetComponent<NetworkObject>().Spawn(true);
         }
 
         if (Asteroid.numAsteroids == 0)